Remove non paginated version
* Also fixes tests
