+++ backend/tests/unit/graphql/test_mutation_update.py
@@ -5,9 +5,7 @@ from neo4j import AsyncSession
 from infrahub.core.branch import Branch
 from infrahub.core.manager import NodeManager
 from infrahub.core.node import Node
-from infrahub.graphql import (
-    generate_graphql_paginated_schema as generate_graphql_schema,
-)
+from infrahub.graphql import generate_graphql_schema
 
 
 async def test_update_simple_object(db, session: AsyncSession, person_john_main: Node, branch: Branch):
